+++ semcore/date-picker/__tests__/index.test.tsx
@@ -333,6 +333,22 @@ describe('DateRangePicker', () => {
     expect(getByText('February 2024')).toBeTruthy();
   });
 
+  test('Should render correctly with empty period', async ({ task, expect }) => {
+    mockDate('2024-01-20T12:00:00.000Z');
+
+    const component = (
+      <DateRangePicker
+        periods={[]}
+        visible
+        disablePortal // only for render popper in test
+        // @ts-ignore
+        __disablePopper // only for render popper in test
+      />
+    );
+
+    await expect(await snapshot(component)).toMatchImageSnapshot(task);
+  });
+
   test('a11y', async () => {
     const { container } = render(
       <DateRangePicker visible disablePortal aria-label='data range picker'>