+++ semcore/utils/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.19.0] - 2024-01-30
+
+### Changed
+
+- options in `a11yEnhance`. `childSelector` could be a function with props from the calling component.
+
+## [4.18.1] - 2024-01-23
+
+### Fixed
+
+- `--intergalactic-chart-grid-border` token for dark theme.
+
 ## [4.18.0] - 2024-01-19
 
 ### Added